requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
boto3==1.26.79
botocore==1.29.79
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-rest-auth==0.9.5
django-silk==5.0.3
django-storages==1.13.2
django-stubs==1.15.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==1.9.0
drf-yasg==1.21.5
exceptiongroup==1.1.0
filelock==3.9.0
frozenlist==1.3.3
gprof2dot==2022.7.29
httplib2==0.21.0
identify==2.5.18
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
multidict==6.0.4
mypy==1.0.1
mypy-extensions==1.0.0
mysqlclient==2.1.1
nested-multipart-parser==1.5.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.7
types-requests==2.28.11.14
types-urllib3==1.26.25.6
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
yarl==1.8.2
